backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.124.0
h11==0.16.0
httptools==0.7.1
httpx==0.27.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.2.1
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
sqlalchemy==2.0.35
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7

# Google Cloud Vertex AI for Imagen image generation
google-cloud-aiplatform==1.38.0
pillow==10.1.0
